isolating documents and anchorable functions from app store to its own store
